refactor(#164): finish annotations/ split — orchestrator.py + facade (P5e)
Move _auto_annotate (the orchestrator) and its helpers (_wrap_rows,
_is_concentric_hole, _concentric_bore_diams, _detect_step_repeat,
_maybe_tabulate_holes) into annotations/orchestrator.py, completing the split of
the 2587-line annotate.py into the annotations/ subpackage. annotate.py is now a
13-line compat facade re-exporting _auto_annotate (+ the two helpers tests
reference), so make_drawing and tests keep importing from draftwright.annotate.

orchestrator.py imports the capability passes (annotations.{sections,turned,pmi,
holes}) and _core/layout — never make_drawing — so the module graph stays a DAG.
The envelope OD/width/depth/height/step dims remain inline in _auto_annotate;
extracting them into annotations/envelope.py is a behaviour-sensitive refactor
(interwoven _dim() calls, not a clean move) deferred as a follow-up.

Pure move: golden gate passes against the committed snapshots with no regeneration;
the full fast tier passes (420); ruff check + format + mypy clean. Adversarially
checked: annotate.py has zero defs (pure facade), orchestrator imports the passes
downward with no upward import, a representative build drives all passes lint-clean.

Closes #164.
